+++ BlueBerry.ToysShop.Web/Models/Visitor.cs
@@ -5,7 +5,7 @@
 		public int Id { get; set; }
 		public string UserId { get; set; }
 		public int ProductId { get; set; }
-		public int ProductRating { get; set; }	
+		public double ProductRating { get; set; }	
 		public string Name { get; set; }
 		public string Comment { get; set; }
 		public DateTime Created { get; set; }
